Makefile: don't inject DESTDIR into the pkgdatadir
In commit 46c4def0733a78ce08702d188e3e1a141fb07316, we added support for
nonstandard PREFIX installations, but DESTDIR was and is never supposed
to be a part of that. While DESTDIR isn't terribly likely to be used
during `make all` invocations, that's no reason to break horribly if it
is used for some reason.