archrelease: actually fail in failure conditions
When svn ls fails due to network timeouts, this currently results in
archrelease deleting all files, then committing this as the changeset.
This causes data loss...

Fix by checking to make sure we actually found files. In the case where
"there are no files" is a legitimate possibility, use a flag value on
errors. If something terrible happened, then we can return failure and
svn's own error messages will provide the rest of the info the user needs.