makechrootpkg: with -n, check if the package failed to install (round 2)
This reimplements the feature initially implemented in commit
be44b9cde15f3228839253c0c0d7d56c124c4e26 andd withdrawn in commit
6e466ca654c9d15e40ae371dbb013170fbdfdb91 by using pacutils instead. If
a package fails to install, we want to make that an error, because
something is certainly seriously wrong with the package, but at the
same time, we also want to allow split packages with conflicts to
succeed, or packages that conflict agaist their own makedepends tree.

This can be performed successfully using makepkg itself, but not with
makepkg --noconfirm, since the prompt default is to simply fail.
pacutils solves this, by allowing the resolution method to be tuned.

This actually turns out to be more useful -- as a side effect of this
change, we now install each split package independently and then namcap
it, before proceeding to the next package (which may cause the previous
package to be uninstalled). This may potentially result in increased
accuracy of some namcap checks which depend on the package being
installed in order to function.