offload-build: correctly handle makepkg's  "-p" argument
This will cause the remote makechrootpkg to try to find a differently
named PKGBUILD, and fail, because we only transferred over the default
one in our --source tarball.

Instead, intercept this argument and pass it to makepkg --source which
will transfer it over to the remote, with the name "PKGBUILD", and use
it as expected. When copying it back over, copy it on top of the right
file.